(Partially) Fixed handling of generic factory methods
If factory method was generic (like `static <E> List<E> of(E e1)` this lead to a wrong signature of the created method since the generics were not correctly resolved.

Quick fix was to check the generic placeholder of the factory against generic class parameters. Note that this might lead to unexpected results if the method uses differently named placeholders than the class itself. Still even in that case, the created class is valid.

Probably more handling of generics is necessary later.